+++ lib/level-zero/event.jl
@@ -50,7 +50,7 @@ Base.unsafe_convert(::Type{ze_event_handle_t}, event::ZeEvent) = event.handle
 signal(event::ZeEvent) = zeEventHostSignal(event)
 append_signal!(list::ZeCommandList, event::ZeEvent) = zeCommandListAppendSignalEvent(list, event)
 
-Base.wait(event::ZeEvent, timeout::Number=typemax(UInt32)) =
+Base.wait(event::ZeEvent, timeout::Number=typemax(UInt64)) =
     zeEventHostSynchronize(event, timeout)
 append_wait!(list::ZeCommandList, events::ZeEvent...) =
     zeCommandListAppendWaitOnEvents(list, length(events), [events...])
